Move types to dedicated types.ts file
