+++ lib/modules/datasource/jsdelivr/types.ts
@@ -1,0 +1,7 @@
+export type JsDelivrPackageType = 'npm' | 'gh';
+
+export interface JsDelivrParsedPackageName {
+  type: JsDelivrPackageType;
+  package: string;
+  asset: string;
+}
